+++ spec/moped/query_spec.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
 
-describe Mongoid::Criteria do
+describe Moped::Query do
   context "scrollable" do
     subject do
-      Feed::Item
+      Feed::Item.collection.find
     end
     it ":scroll" do
       subject.should.respond_to? :scroll
@@ -11,7 +11,7 @@ describe Mongoid::Criteria do
   end
   context "with multiple sort fields" do
     subject do
-      Feed::Item.desc(:name).asc(:value)
+      Feed::Item.collection.find.sort(name: 1, value: -1)
     end
     it "raises Mongoid::Scroll::Errors::MultipleSortFieldsError" do
       expect { subject.scroll }.to raise_error Mongoid::Scroll::Errors::MultipleSortFieldsError,
@@ -20,10 +20,10 @@ describe Mongoid::Criteria do
   end
   context "with no sort" do
     subject do
-      Feed::Item.all
+      Feed::Item.collection.find
     end
     it "adds a default sort by _id" do
-      subject.scroll.options[:sort].should == { "_id" => 1 }
+      subject.scroll.operation.selector["$orderby"].should == { "_id" => 1 }
     end
   end
   context "with data" do
@@ -41,45 +41,45 @@ describe Mongoid::Criteria do
     context "default" do
       it "scrolls all" do
         records = []
-        Feed::Item.all.scroll do |record, next_cursor|
+        Feed::Item.collection.find.scroll do |record, next_cursor|
           records << record
         end
         records.size.should == 10
-        records.should eq Feed::Item.all.to_a
+        records.should eq Feed::Item.collection.find.to_a
       end
     end
     { a_string: String, a_integer: Integer, a_date: Date, a_datetime: DateTime }.each_pair do |field_name, field_type|
       context field_type do
         it "scrolls all with a block" do
           records = []
-          Feed::Item.asc(field_name).scroll do |record, next_cursor|
+          Feed::Item.collection.find.sort(field_name => 1).scroll do |record, next_cursor|
             records << record
           end
           records.size.should == 10
-          records.should eq Feed::Item.all.to_a
+          records.should eq Feed::Item.collection.find.to_a
         end
         it "scrolls all with a break" do
           records = []
           cursor = nil
-          Feed::Item.asc(field_name).limit(5).scroll do |record, next_cursor|
+          Feed::Item.collection.find.sort(field_name => 1).limit(5).scroll do |record, next_cursor|
             records << record
             cursor = next_cursor
           end
           records.size.should == 5
-          Feed::Item.asc(field_name).scroll(cursor) do |record, next_cursor|
+          Feed::Item.collection.find.sort(field_name => 1).scroll(cursor) do |record, next_cursor|
             records << record
             cursor = next_cursor
           end
           records.size.should == 10
-          records.should eq Feed::Item.all.to_a
+          records.should eq Feed::Item.collection.find.to_a
         end
         it "scrolls in descending order" do
           records = []
-          Feed::Item.desc(field_name).limit(3).scroll do |record, next_cursor|
+          Feed::Item.collection.find.sort(field_name => -1).limit(3).scroll(nil, { field_type: field_type }) do |record, next_cursor|
             records << record
           end
           records.size.should == 3
-          records.should eq Feed::Item.desc(field_name).limit(3).to_a
+          records.should eq Feed::Item.collection.find.sort(field_name => -1).limit(3).to_a
         end
         it "map" do
           record = Feed::Item.desc(field_name).limit(3).scroll.map { |record, cursor| record }.last
@@ -100,16 +100,16 @@ describe Mongoid::Criteria do
     it "scrolls" do
       records = []
       cursor = nil
-      Feed::Item.desc(:a_integer).limit(2).scroll do |record, next_cursor|
+      Feed::Item.collection.find.sort(a_integer: -1).limit(2).scroll do |record, next_cursor|
         records << record
         cursor = next_cursor
       end
       records.size.should == 2
-      Feed::Item.desc(:a_integer).scroll(cursor) do |record, next_cursor|
+      Feed::Item.collection.find.sort(a_integer: -1).scroll(cursor) do |record, next_cursor|
         records << record
       end
       records.size.should == 3
-      records.should eq Feed::Item.all.to_a
+      records.should eq Feed::Item.collection.find.to_a
     end
   end
 end