+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Next Release
+============
+
+* Extended `Moped::Query` with `scroll` - [@dblock](https://github.com/dblock).
+* `Mongoid::Scroll::Cursor.from_record` can now be called with either a Mongoid field or `field_type` and `field_name` in the `options` hash - [@dblock](https://github.com/dblock).
+
 0.1.0 (2/14/2013)
 =================